Update Asterisk configuration with specific port range
Like that, Asterisk will now use a specific range of ports stated in the rtp configuration file.
UDP ports between 10.000 and 15.000 in order to make correctly audio channels and so on.
Still need to be checked because problem on the VPS side (I hope, no changes on the Docker side).